refactor: split features/workflow/engine.go into cohesive modules (Issue #449)
features/workflow/engine.go had grown to 3,110 lines with 75 methods on a single
Engine struct handling unrelated concerns: condition evaluation, HTTP/API calls,
loop execution, synchronization primitives, nested workflow composition, fan-out/fan-in,
error workflows, and composite steps. Despite the monolithic file, the 16 test files
were already well-separated by domain, confirming these are distinct concerns.

Dependency analysis identified leaf-node method groups (zero calls to other Engine
methods) as ideal split candidates. Fan-out was kept in core since executeFanOutStep
calls executeStep, which would create a circular reference.

Split into 5 new files by extracting leaf nodes first:

- engine_conditions.go (366 lines): evaluateCondition + 17 helper methods for
  condition/expression evaluation. Zero external Engine method dependencies.
  Tests: conditional_test.go (already isolated).

- engine_http.go (156 lines): executeHTTPStep, executeAPIStep, executeWebhookStep,
  executeDelayStep. Uses only httpClient and providerRegistry (leaf nodes).
  Tests: http_test.go.

- engine_sync.go (255 lines): executeBarrierStep, executeSemaphoreStep,
  executeLockStep, executeWaitGroupStep. Uses only syncManager (leaf nodes).
  Tests: sync_test.go.

- engine_loops.go (403 lines): executeForStep, executeWhileStep, executeForeachStep
  + resolveLoopValue, convertToSlice, convertToInt, parseInteger utilities.
  Tests: loop_test.go, loop_break_test.go.

- engine_composition.go (829 lines): executeWorkflowStep, executeErrorWorkflowStep,
  executeCompositeStep + nested/async helpers, component execution methods,
  executeFanInStep + fanIn helpers. Tests: nested_workflow_test.go, error_workflow_test.go.

engine.go (core, 1,146 lines): execution lifecycle, step dispatch, state management,
debug integration, try-catch, fan-out/fan-in collection.

All files remain in package workflow — no circular imports possible.
All 16 existing test files pass with zero test modifications.
No public API surface changes; pure file reorganization.

Changes:
- Trim engine.go from 3,110 to 1,146 lines (-63%)
- Create engine_conditions.go (366 lines, 18 condition/expression methods)
- Create engine_http.go (156 lines, 4 HTTP/API methods)
- Create engine_sync.go (255 lines, 4 synchronization methods)
- Create engine_loops.go (403 lines, 7 loop methods + utilities)
- Create engine_composition.go (829 lines, nested workflows + fan-in + composite)

Fixes #449
